+++ src/lib.rs
@@ -1,5 +1,8 @@
 #![doc = include_str!("../README.md")]
+#![cfg_attr(feature = "no_std", no_std)]
 #![cfg_attr(feature = "nightly", feature(float_gamma))]
+
+extern crate alloc;
 
 mod binary;
 mod count;