+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Added `Error` enum
+- Added experimental support for `no_std`
 - Added experimental `nightly` feature
 
 ## 0.2.0 (2024-07-10)